pyproject.toml
[tool.poetry]
name = "deliveryapi"
version = "0.1.0"
description = ""
authors = ["acronicsM <91975262+acronicsM@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
aiosqlite = "^0.19.0"
pydantic-settings = "^2.1.0"
alembic = "^1.12.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.1"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
